// src/app/common/enum/kyc/kyc.actividad.enum.ts
export enum EKycActividad {
  FIN = 0,
  NUEVO_FOLIO = 1,
  SOLICITUD = 2,
  CARGA_DOCUMENTAL = 3,
  VALIDACION_DIGITAL = 4,
  FIRMA_CLIENTE = 5,
  FIRMA_EJECUTIVO = 6,
  VALIDACION_FIRMAS = 7,
  CONFIRMACION_ENTREGA = 8,
  CONTACTO_TELEFONICO = 9,
  CONTACTO_ASEGURADORA = 10,
  SUSCRIPCION = 11,
  INFORMACION_ACLARATORIA = 12,
  CONSULTA = 12,
  CONSULTA_EXCEPCION = 13,
  GENERACION_PAGO = 14,
  CONFIRMA_PAGO = 15,
  GENERACION_FINIQUITO = 16,
  FIRMA_FINIQUITO = 17,
  RECOLECCION_FISICOS = 18,
  VALIDACION_ORIGINALES = 19,
  RECEPCION_ORIGINALES = 20,
  ESGRACIAS = 21,
  VALIDACION_AFIANZADORA = 29,

}